requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
httpx==0.27.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
